check cancel status, revise fix to #60
- Ignore spawning mobs if initial spawn was canceled
- getDrops returns a new list of drops said block would drop when
broken. CB does not use this list when creating said drops on break,
so... clearing this list pretty much does nothing.

Given the nature of "breaking" the block without dropping anything must
be performed differently due to each event "breaking" said skull, we'll
defer this task to the TemporaryBlockHandler instead of trying to do it
in a TemporaryBlockEvent listener (and is probably the "right" way to do
it anyways).