revert the disable module config
Either the modules have to be enabled in a certain order, or they're
dependent on each other.

Also, if these are impacting performance, we need to go about fixing
these instead of just blanket-disabling them. Debug could probably be
disabled, although it seems a bit involved to do so(?) - or it could
just be simply removed.

The tutorial module should have an individual option for storing
persistent data. Might be worth creating a global config option for this
(albeit it'd be the only option in there for the moment...)