Wow, this really is a deprecated abstraction jungle
If it took this much work to do 1.13 support using deprecated APIs, just
imagine how much more time would be needed to convert to non-deprecated
APIs...

Anyways, some things have been temporarily removed or half-converted.
All relevant areas are marked with //TODO: 1.13. While a lot of it was
just converting Material names, there are quite a number of methods that
do other deprecated things, some of which were actually removed in the
API.

The only deleted things are MaterialHelper and its test, as it's only
used to give the plural form of material types when printing the item
types lost on death.