Use placeevent#getBlockAgainst instead of checking for attachment material
Since only wall torches "attach" to blocks (and apparently the Torch
constructor doesn't do this check properly?), use getBlockAgainst
instead to check for the attachment.

The torch is not guaranteed to be placed/attached to the getBlockAgainst
block though, but should stop most cases, as the idea is to prevent
landscapes from being littered with torches moreso then preventing
torches from being attached to any soft material.

remove debug